+++ CHANGELOG.md
@@ -8,6 +8,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 - Moved classes with extension methods to `SharpX.Extensions` namespace.
+- Moved `StringUtil::Generate` to new `Strings` class.
+- Moved `CharExtensions::Replicate` as `Strings::ReplicateChar`
+- Moved `CharExtensions::Replicate` to `StringExtensions`.
+- Removed `CharExtensions` class.
+- All `StringExtensions` methods are implemented in `Strings` class.
 
 ## [1.0.3] - 2021-11-20